+++ src/interface/index.ts
@@ -1,0 +1,5 @@
+export interface IStateRequest<T> {
+  isLoading: boolean;
+  data?: null | T[];
+  error?: null | T[];
+}
