+++ src/pages/Pokedex/index.tsx
@@ -1,6 +1,6 @@
 import { A } from 'hookrouter';
 import React, { useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import Heading from '../../components/Heading';
 import Layout from '../../components/Layout';
 import PokemonCard from '../../components/PokemonCard';
@@ -8,7 +8,7 @@ import useData from '../../hook/useData';
 import useDebounse from '../../hook/useDebounse';
 import { IPokemonsResponse } from '../../interface/pokemons';
 import IQuery from '../../interface/query';
-import { getTypesAction } from '../../store/pokemons';
+import { getPokemonsTypes, getPokemonsTypesLoading, getTypesAction } from '../../store/pokemons';
 import s from './style.module.scss';
 
 export interface IPokedexPage {
@@ -17,6 +17,8 @@ export interface IPokedexPage {
 
 const PokedexPage: React.FC<IPokedexPage> = ({ id }: IPokedexPage) => {
   const dispatch = useDispatch();
+  const types = useSelector(getPokemonsTypes);
+  const isTypesLoading = useSelector(getPokemonsTypesLoading);
   const [searchValue, setSearchValue] = useState('');
   const [query, setQuery] = useState<IQuery>({
     limit: 12,
@@ -46,6 +48,13 @@ const PokedexPage: React.FC<IPokedexPage> = ({ id }: IPokedexPage) => {
         <div className={s.inputStyle}>
           <Heading level="h3">Please input Pokemon Name</Heading>
           <input key="editor1" type="text" value={searchValue} onChange={handleSearchChange} />
+          <div>
+            {isTypesLoading ? (
+              <Heading className={s.message}>Is Loading...</Heading>
+            ) : (
+              <Heading level="h5">{types?.join(' ')}</Heading>
+            )}
+          </div>
           <p>
             {data?.total ? (
               <>
