+++ src/store/index.tsx
@@ -1,7 +1,10 @@
 import { combineReducers } from 'redux';
 import app from './app';
-import pokemons from './pokemons';
+import pokemons, { IPokemonsInitialState } from './pokemons';
 
+export interface IInitialState {
+  pokemons: IPokemonsInitialState;
+}
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 const createRootReducer = () => combineReducers({ app, pokemons });
